run npm pack twice (#32)
Run the npm pack to build the package and then again in "--dry" mode to fetch the name of the package. 

We do this because npm updated versions and now "npm pack" outputs garbage to stdout, messing up with artefact name capture. So we use one run to actually build and output the logs, and the other run without output for just fetching the artefact name.

Signed-off-by: Josh Leyshon <josh.leyshon@zepben.com>